PE-0119: Takes too long with erorrs.

+++ Demo.ProjectEuler/Demo.ProjectEuler.Core/NumberUtil.cs
@@ -1,14 +1,15 @@
 ﻿using System.Collections.Generic;
+using System.Numerics;
 
 namespace Demo.ProjectEuler.Core
 {
 	public class NumberUtil
 	{
-		public IEnumerable<int> ToReverseSequence(long number)
+		public IEnumerable<long> ToReverseSequence(BigInteger number)
 		{
 			while (number > 0)
 			{
-				int digit = (int)number % 10;
+				long digit = (long) number % 10;
 				yield return digit;
 
 				number /= 10;
