+++ Demo.ProjectEuler/Demo.ProjectEuler.Tests/LibraryTests/NumberUtilTest.cs
@@ -1,4 +1,5 @@
 ﻿using System.Linq;
+using System.Numerics;
 using Demo.ProjectEuler.Core;
 using Demo.ProjectEuler.Tests.Core;
 using Xunit;
@@ -21,9 +22,9 @@ namespace Demo.ProjectEuler.Tests.LibraryTests
 		[InlineData(2234, new [] {2,2,3,4})]
 		[InlineData(1234567890, new [] {1,2,3,4,5,6,7,8,9,0})]
 		[InlineData(9876, new [] {9,8,7,6})]
-		public void TestNumberToSequence(long input, int[] expected)
+		public void TestNumberToSequence(long input, long[] expected)
 		{
-			int[] actual = _sut.ToReverseSequence(input).Reverse().ToArray();
+			long[] actual = _sut.ToReverseSequence(input).Reverse().ToArray();
 
 			Assert.True(expected.SequenceEqual(actual));
 		}
